feat: add EntityImmutable class.
